prod

+++ database/migrations/0000_00_00_000008_create_villages_table.php
@@ -22,18 +22,19 @@ return new class extends Migration
             $table->integer( "x" )->default( 0 );
             $table->integer( "y" )->default( 0 );
             $table->string( "map_sector" )->default( 0 );
-            $table->integer( "points" )->default( 21 );
-            $table->double( "loyalty" )->default( 100 );
+
+            $table->integer( "points"  )->default(  21 );
+            $table->integer( "loyalty" )->default( 100 );
 
             $table->double( "prod_wood" )->default( 0 );
             $table->double( "prod_clay" )->default( 0 );
             $table->double( "prod_iron" )->default( 0 );
 
-            $table->integer( "pop" )->default( 5 );
+            $table->integer( "pop"                 )->default(   5 );
 
-            $table->double( "stored_wood" )->default( 600 );
-            $table->double( "stored_clay" )->default( 600 );
-            $table->double( "stored_iron" )->default( 600 );
+            $table->integer( "stored_wood"         )->default( 600 );
+            $table->integer( "stored_clay"         )->default( 600 );
+            $table->integer( "stored_iron"         )->default( 600 );
 
             $table->integer( "building_main"       )->default( 1 );
             $table->integer( "building_barracks"   )->default( 0 );
